Fix timestamp overflow: Remove uint32_t cast for nTime
- Remove static_cast<uint32_t> from nTime assignment
- TritiumBlock::nTime is uint64_t, not uint32_t
- Prevents overflow after year 2106 when Unix timestamps exceed 32 bits
- Fixed in both stateless_block_utility.cpp and create.cpp for consistency
- std::max result is uint64_t, directly assigned to uint64_t field

+++ src/TAO/Ledger/create.cpp
@@ -343,7 +343,7 @@ namespace TAO::Ledger
         block.nHeight       = tStateBest.nHeight + 1;
         block.nBits         = GetNextTargetRequired(tStateBest, nChannel, false);
         block.nNonce        = 1;
-        block.nTime         = static_cast<uint32_t>(std::max(tStateBest.GetBlockTime() + 1, runtime::unifiedtimestamp()));
+        block.nTime         = std::max(tStateBest.GetBlockTime() + 1, runtime::unifiedtimestamp());
     }
